server/requirements.txt
amqp==2.4.2
anyjson==0.3.3
attrs==19.1.0
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.2
certifi==2019.3.9
chardet==3.0.4
constantly==15.1.0
Django==2.1.7
django-celery-beat==1.4.0
django-cors-headers==2.5.2
django-timezone-field==3.0
djangorestframework==3.9.1
djangorestframework-stubs==0.3.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
kombu==4.3.0
Pillow==5.4.1
psycopg2==2.7.7
PyHamcrest==1.9.0
python-crontab==2.3.6
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
redis==3.2.1
requests==2.21.0
six==1.12.0
Twisted==18.9.0
urllib3==1.24.1
vine==1.3.0
zope.interface==4.6.0
